+++ components/eamxx/src/share/CMakeLists.txt
@@ -27,6 +27,7 @@ set(SHARE_SRC
   grid/remap/horiz_interp_remapper_data.cpp
   grid/remap/refining_remapper_p2p.cpp
   grid/remap/vertical_remapper.cpp
+  iop/intensive_observation_period.cpp
   property_checks/property_check.cpp
   property_checks/field_nan_check.cpp
   property_checks/field_within_interval_check.cpp
